Move file lookup code out of parser_common
This is to avoid adding extra dependency on RBT.
