+++ multibody/parsers/sdf_parser.cc
@@ -8,7 +8,6 @@
 #include <utility>
 #include <vector>
 
-#include <spruce.hh>
 #include <tinyxml2.h>
 
 #include "drake/common/eigen_types.h"
@@ -17,6 +16,7 @@
 #include "drake/multibody/joints/floating_base_types.h"
 #include "drake/multibody/parsers/model_instance_id_table.h"
 #include "drake/multibody/parsers/parser_common.h"
+#include "drake/multibody/parsers/parser_path_utils.h"
 #include "drake/multibody/parsers/xml_util.h"
 #include "drake/multibody/rigid_body_plant/compliant_material.h"
 #include "drake/multibody/rigid_body_tree.h"
@@ -997,4 +997,3 @@ ModelInstanceIdTable AddModelInstancesFromSdfStringSearchingInRosPackages(
 }  // namespace sdf
 }  // namespace parsers
 }  // namespace drake
-
