+++ multibody/parsers/urdf_parser.cc
@@ -14,6 +14,7 @@
 #include "drake/multibody/joints/floating_base_types.h"
 #include "drake/multibody/material_map.h"
 #include "drake/multibody/parsers/model_instance_id_table.h"
+#include "drake/multibody/parsers/parser_path_utils.h"
 #include "drake/multibody/parsers/xml_util.h"
 #include "drake/multibody/rigid_body_plant/compliant_material.h"
